# コミットメッセージ
[refactor] ホームページコンポーネントの動的インポート対応

## 詳細内容
- `viewer/src/app/page.tsx` ページコンポーネントを簡素化し、動的インポートを導入
  - Suspenseとdynamic importを使用してHomePageを遅延読み込み
- `viewer/src/components/home-page.tsx` 既存のページロジックを新コンポーネントに移行
  - WebSocket接続ロジックとUI構造をそのまま移動
